Don't accidentally extend implicit `eval` function
This package uses the name `eval`, which is ordinarily reserved for
the implicitly provided `eval` function provided by the core system.
Adding methods to this generic function worked accidentally due to
the way the implementation works, but is probably neither what you
want nor guarateed to keep working (e.g. https://github.com/JuliaLang/julia/pull/55949
would break it if merged). To address the issue, make `Fuzzy` a
baremodule to avoid implicitly creating the `include`/`eval` names
and then add back explicit imports of Base and a definition of `include`.

This way, `Fuzzy.eval` is completely decoupled from the core notion of
`eval`.